+++ src/components/features/coupon/index.ts
@@ -1,0 +1,6 @@
+export * from './api';
+export { default as CouponItem } from './couponItem/CouponItem';
+export { default as CouponModal } from './couponModal/CouponModal';
+export * from './hooks';
+export * from './models';
+export * from './utils';
