+++ src/components/features/coupon/models/coupon.ts
@@ -1,7 +1,9 @@
-import { CartItemType } from '../../cart/types';
-import { calculateDeliveryFee } from '../../cart/utils/calculateDeliveryFee';
-import { calculateOrderPrice } from '../../cart/utils/cartCalculations';
-import { convertTimeToSecond } from '../utils/time';
+import {
+  CartItemType,
+  calculateDeliveryFee,
+  calculateOrderPrice,
+} from '../../cart';
+import { convertTimeToSecond } from '../utils';
 import { CouponType } from './coupon.types';
 import {
   createCoupon,
