+++ src/components/features/coupon/models/index.ts
@@ -1,0 +1,2 @@
+export { default as Coupon } from './coupon';
+export * from './coupon.types';
