+++ src/components/features/payment/index.ts
@@ -1,0 +1,1 @@
+export { default as PaymentCheckContents } from './PaymentCheckContents/PaymentCheckContents';
